Move sealed module from app to plugin
