+++ crates/bevy_app/src/app.rs
@@ -14,7 +14,7 @@ use std::fmt::Debug;
 #[cfg(feature = "trace")]
 use bevy_utils::tracing::info_span;
 
-use self::sealed::{IntoPlugin, IntoPluginGroup};
+use super::sealed::{IntoPlugin, IntoPluginGroup};
 bevy_utils::define_label!(
     /// A strongly-typed class of labels used to identify an [`App`].
     AppLabel,
@@ -1002,86 +1002,6 @@ impl App {
 
         self
     }
-}
-
-mod sealed {
-    use bevy_ecs::all_tuples;
-
-    use crate::{App, Plugin, PluginGroup, PluginGroupBuilder};
-
-    pub trait IntoPlugin<Marker> {
-        type Plugin: Plugin;
-        fn into_plugin(self, app: &mut App) -> Self::Plugin;
-    }
-
-    pub trait IntoPluginGroup<Marker>: IntoPluginGroupBuilder<Marker> {}
-
-    pub trait IntoPluginGroupBuilder<Marker> {
-        fn into_plugin_group_builder(self, app: &mut App) -> PluginGroupBuilder;
-    }
-
-    pub struct IsPlugin;
-    pub struct IsPluginGroup;
-    pub struct IsFunction;
-
-    impl<P: Plugin> IntoPlugin<IsPlugin> for P {
-        type Plugin = Self;
-        fn into_plugin(self, _: &mut App) -> Self {
-            self
-        }
-    }
-
-    impl<P: Plugin> IntoPluginGroupBuilder<IsPlugin> for P {
-        fn into_plugin_group_builder(self, _: &mut App) -> PluginGroupBuilder {
-            PluginGroupBuilder::from_plugin(self)
-        }
-    }
-
-    impl<P: PluginGroup> IntoPluginGroupBuilder<IsPluginGroup> for P {
-        fn into_plugin_group_builder(self, _: &mut App) -> PluginGroupBuilder {
-            self.build()
-        }
-    }
-
-    impl<P: PluginGroup> IntoPluginGroup<IsPluginGroup> for P {}
-
-    impl<F: FnOnce(&mut App) -> P, P: Plugin> IntoPlugin<IsFunction> for F {
-        type Plugin = P;
-
-        fn into_plugin(self, app: &mut App) -> Self::Plugin {
-            self(app)
-        }
-    }
-
-    impl<F: FnOnce(&mut App) -> PG, PG: PluginGroup> IntoPluginGroupBuilder<IsFunction> for F {
-        fn into_plugin_group_builder(self, app: &mut App) -> PluginGroupBuilder {
-            self(app).build()
-        }
-    }
-
-    impl<F: FnOnce(&mut App) -> PG, PG: PluginGroup> IntoPluginGroup<IsFunction> for F {}
-
-    macro_rules! impl_plugin_collection {
-        ($(($param: ident, $plugins: ident)),*) => {
-            impl<$($param, $plugins),*> IntoPluginGroupBuilder<($($param,)*)> for ($($plugins,)*)
-            where
-                $($plugins: IntoPluginGroupBuilder<$param>),*
-            {
-                #[allow(non_snake_case, unused_variables)]
-                fn into_plugin_group_builder(self, app: &mut App) -> PluginGroupBuilder {
-                    let ($($plugins,)*) = self;
-                    PluginGroupBuilder::merge(vec![$($plugins.into_plugin_group_builder(app),)*])
-                }
-            }
-
-            impl<$($param, $plugins),*> IntoPluginGroup<($($param,)*)> for ($($plugins,)*)
-            where
-                $($plugins: IntoPluginGroupBuilder<$param>),*
-            {}
-        }
-    }
-
-    all_tuples!(impl_plugin_collection, 0, 15, P, S);
 }
 
 fn run_once(mut app: App) {
